refactor: adiciona express e move o front pra uma pasta frontend
