+++ poker-planning-v3/package.json
@@ -14,9 +14,11 @@
     "@tailwindcss/vite": "^4.1.10",
     "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
+    "express": "^5.1.0",
     "lucide-react": "^0.516.0",
     "react": "^19.1.0",
     "react-dom": "^19.1.0",
+    "socket.io": "^4.8.1",
     "tailwind-merge": "^3.3.1",
     "tailwindcss": "^4.1.10"
   },
