+++ poker-planning-v3/vite.config.ts
@@ -6,6 +6,7 @@ import tailwindcss from "@tailwindcss/vite";
 // https://vite.dev/config/
 export default defineConfig({
   plugins: [react(), tailwindcss()],
+  root: "src/frontend",
   server: {
     host: "0.0.0.0",
     port: 5173,
